Spawn respawn timer on score, despawn on respawn
Previously, I had a struct with a timer and a boolean indicating whether
or not the timer was active. The timer would be set to active on score,
and set to inactive (and reset) once the ball was respawed.

This felt terrible for a number of reasons:

1. The timer existed even when the ball was not waiting to respawn.
2. Maintaining the timer's state felt really dirty, as its active field
   was modified in multiple places. Not really intuitive how it worked.
3. When I try to explain why the active field exists or what it does to
   myself, it just sounds dumb as shit.
4. Just ew.

Now, I simply spawn a timer as a resource when the ball is deleted, and
delete the timer when the ball is respawned. I think this will be
reworked in the future.